Updates paper-button to match material design guidelines.
Several bugs are addressed in this commit.  At default button font size
of 14px, the new button will have the following dimensions:
* Min-width of 64px
* Ripple effect only covers the raised portion
* Touch height of 48px
* Button minimum height of 36px
* Shadow visible on all sides of raised button.

The new button has 4px of external margin built-in that (unfortunately)
are also part of the touch target.  However, this appears to be
necessary for paper-shadow to correctly render.